Delivery

+++ Assets/_Scripts/Kitchen/Delivery/DeliveryCounterView.cs
@@ -1,4 +1,5 @@
 ﻿using DG.Tweening;
+using System;
 using System.Collections.Generic;
 using UnityEngine;
 
@@ -14,10 +15,10 @@ public class DeliveryCounterView : CounterView
 
     protected override void Initialize()
     {
-        Model.OnPlateDelivered += DeliveryCounter_OnPlateDelivered;
+        Model.OnItemChanged += DeliveryCounter_OnPlateDelivered;
         Model.OnPlateReleased += DeliveryCounter_OnPlateReleased;
 
-        _slideDuration = Model.DeliveryDelay;
+        _slideDuration = Model.DeliveryDelay * 2f;
     }
 
     protected override void SetupComponents()
@@ -25,8 +26,12 @@ public class DeliveryCounterView : CounterView
         counterEnd = counterEnd != null ? counterEnd : transform;
     }
 
-    private void DeliveryCounter_OnPlateDelivered(PlateObjectController po)
+    private void DeliveryCounter_OnPlateDelivered(KitchenObjectController ko)
     {
+        var po = ko as PlateObjectController;
+        if (po == null)
+            return;
+
         if (activeTweens.ContainsKey(po))
         {
             activeTweens[po].Kill();
@@ -38,10 +43,11 @@ public class DeliveryCounterView : CounterView
                          counterEnd.position,
                          _slideDuration)
             .SetEase(Ease.Linear)
+            .SetDelay(0.2f)
             .OnComplete(() =>
             {
                 activeTweens.Remove(po);
-                Destroy(po.gameObject);
+                po.DestroySelf();
             });
 
         activeTweens[po] = deliveryTween;
